Moved cowswap assets to custom/assets/cow-swap folder
